pyproject.toml
[tool.poetry]
name = "my-clean-DI"
version = "0.1.0"
description = ""
authors = ["movassaghi6"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
fastapi-utils = "^0.2.1"
uvicorn = "^0.18.2"
gunicorn = "^20.1.0"
python-dotenv = "^0.20.0"
mongoengine = "^0.24.1"
loguru = "^0.6.0"
requests = "^2.28.1"
anyio = "^4.4.0"
httpx = "^0.27.0"
pytest-anyio = "^0.0.0"
pytest-mongodb = "^2.4.0"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"
python-multipart = "^0.0.9"

[tool.poetry.dev-dependencies]
essential-generators = "^1.0"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
pytest-xdist = "^2.3"
asgi-lifespan = "^1.0.1"
coverage = "^6.4.1"
coverage-badge = "1.1.0"
pytest-dotenv = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
